fix (index.js) Fix hot reloading of new submitted products.
Whenever you added a new product, the submission update did not happened automatically to show new db changes. Had to add window.location.reload() to several sections in order to be able to see latest updates as soon as they were submitted into the database.
Transactions were as well not being registered whenever a new product was sold. This required the creation of a new transaction inside the sell resolver in order to add it to its respective collection.

No breaking changes.