section (index.js) Add content to dashboard-resources
Dashboard-resources did not have any content available. Now it is possible to upload links or files to the resources section. Had to do an AWS integration to submit files to an S3 bucket. Credentials are currently under my account, but a migration to HQT accounts is intended.

No breaking changes.
**New .env variables will be added manually in the server.